refactor: moved tests to FinishedImportingIT files
