+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/opensearch/OpensearchFinishedImportingIT.java
@@ -36,6 +36,7 @@ import java.util.Arrays;
 import org.awaitility.Awaitility;
 import org.junit.Before;
 import org.junit.Test;
+import org.opensearch.client.opensearch.core.SearchRequest;
 import org.opensearch.client.opensearch.core.SearchRequest.Builder;
 import org.opensearch.client.opensearch.core.search.Hit;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -252,6 +253,29 @@ public class OpensearchFinishedImportingIT extends OperateZeebeAbstractIT {
 
     // then
     assertImportPositionMatchesRecord(record, ImportValueType.PROCESS_INSTANCE, 1);
+  }
+
+  @Test
+  public void shouldWriteDefaultEmptyDefaultImportPositionDocumentsOnRecordReaderStart() {
+    recordsReaderHolder.getAllRecordsReaders().stream()
+        .map(OpensearchRecordsReader.class::cast)
+        .forEach(OpensearchRecordsReader::postConstruct);
+
+    Awaitility.await()
+        .atMost(Duration.ofSeconds(30))
+        .until(
+            () -> {
+              final var searchRequestBuilder =
+                  new SearchRequest.Builder()
+                      .size(100)
+                      .index(importPositionIndex.getFullQualifiedName());
+              final var documents =
+                  osClient.doc().search(searchRequestBuilder, ImportPositionEntity.class);
+
+              // all initial import position documents created for each record reader
+              return documents.hits().hits().size()
+                  == recordsReaderHolder.getAllRecordsReaders().size();
+            });
   }
 
   private void assertImportPositionMatchesRecord(
